+++ src/duckpd/__init__.py
@@ -3,8 +3,8 @@
 from importlib.metadata import version
 
 from duckpd.frame import DataFrame
-from duckpd.groupby import DataFrameGroupBy
-from duckpd.io import from_arrow, from_pandas, read_parquet
+from duckpd.groupby import DataFrameGroupBy, SeriesGroupBy
+from duckpd.io import concat, from_arrow, from_pandas, read_parquet
 from duckpd.series import Series
 from duckpd.session import Session, connect
 
@@ -14,8 +14,10 @@ __all__ = [
     "DataFrame",
     "DataFrameGroupBy",
     "Series",
+    "SeriesGroupBy",
     "Session",
     "__version__",
+    "concat",
     "connect",
     "from_arrow",
     "from_pandas",